RavenDB-24887 - make AiAgentActionRequest.Type internal, move HandleSubAgentCalls, BuildAgentRequest, GetOrAddSubAgentsActionResponses, CreateAgentRequest, _childUserCalls to other partial class, Move back ExecuteMultiAgentAndQueryRequestsAsync method
